+++ lib/src/context.dart
@@ -1,4 +1,5 @@
 import 'dart:io';
+import 'dart:mirrors';
 
 import 'http_error.dart';
 
@@ -27,8 +28,35 @@ class Context {
   ContentType? get contentType => headers.contentType;
   set contentType(ContentType? type) => headers.contentType = type;
 
+  /// Set Response Header key=value
+  void set(String key, String value) {
+    headers.set(key, value);
+  }
+
   /// Throw HttpError with status code, message and stackTrace
   void throwError(int status, {String? message, StackTrace? stackTrace}) {
     throw HttpError(status, message: message, stackTrace: stackTrace);
+  }
+
+  /// create Context or custom Context instance
+  static dynamic createInstance(Type type,
+      {Symbol? constructor,
+      List? arguments,
+      Map<Symbol, dynamic>? namedArguments}) {
+    constructor ??= const Symbol('');
+    arguments ??= const [];
+
+    var typeMirror = reflectType(type);
+    if (typeMirror is ClassMirror) {
+      if (namedArguments != null) {
+        return typeMirror
+            .newInstance(constructor, arguments, namedArguments)
+            .reflectee;
+      } else {
+        return typeMirror.newInstance(constructor, arguments).reflectee;
+      }
+    } else {
+      throw ArgumentError("Cannot create the instance of the type '$type'.");
+    }
   }
 }