+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Make ```createInstance``` public static in ```Context```
+- Add method ```set(String key, String value)``` to ```Context``` as symlink to ```headers.set```
+- Add the link to [dia_router](http://https://github.com/unger1984/dia_router)
+
 ## 0.0.1
 
 - Initial version